requirements.txt
python-socketio==5.6.0
websocket-client